requirements.txt
numpy==1.17.0
jupyter==1.0.0
https://download.pytorch.org/whl/cu100/torch-1.1.0-cp36-cp36m-linux_x86_64.whl
https://download.pytorch.org/whl/cu100/torchvision-0.3.0-cp36-cp36m-linux_x86_64.whl
torchtext==0.4.0
pytorch-transformers==1.2.0
